Update on Admin login
Here I added onto my log in code for admin that if a field is empty access is denied and log in is now through the log in button. My first commit was on the main branch as it was the very first commit, this one is a small change and is on my personal branch.